move transcription related code into standalone composable
