+++ packages/stage-layouts/src/components/Widgets/ChatArea.vue
@@ -9,19 +9,20 @@ import { useAudioContext } from '@proj-airi/stage-ui/stores/audio'
 import { useChatOrchestratorStore } from '@proj-airi/stage-ui/stores/chat'
 import { useChatSessionStore } from '@proj-airi/stage-ui/stores/chat/session-store'
 import { useConsciousnessStore } from '@proj-airi/stage-ui/stores/modules/consciousness'
-import { useHearingSpeechInputPipeline, useHearingStore } from '@proj-airi/stage-ui/stores/modules/hearing'
 import { useProvidersStore } from '@proj-airi/stage-ui/stores/providers'
 import { useSettings, useSettingsAudioDevice } from '@proj-airi/stage-ui/stores/settings'
 import { BasicTextarea, FieldCombobox } from '@proj-airi/ui'
-import { until, useLocalStorage } from '@vueuse/core'
+import { useLocalStorage } from '@vueuse/core'
 import { storeToRefs } from 'pinia'
 import { DropdownMenuContent, DropdownMenuItem, DropdownMenuPortal, DropdownMenuRoot, DropdownMenuTrigger, PopoverContent, PopoverRoot, PopoverTrigger } from 'reka-ui'
-import { computed, nextTick, onUnmounted, ref, watch } from 'vue'
+import { computed, onUnmounted, ref, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
 
 import IndicatorMicVolume from './IndicatorMicVolume.vue'
 
-const messageInput = ref('')
+import { useTranscriptions } from '../../composables/use-transcriptions'
+
+const messageInput = ref<string>('')
 const hearingPopoverOpen = ref(false)
 const sessionsDrawerOpen = ref(false)
 const isComposing = ref(false)
@@ -37,7 +38,7 @@ const providersStore = useProvidersStore()
 const { activeProvider, activeModel } = storeToRefs(useConsciousnessStore())
 const { themeColorsHueDynamic } = storeToRefs(useSettings())
 
-const { askPermission, startStream } = useSettingsAudioDevice()
+const { askPermission } = useSettingsAudioDevice()
 const { enabled, selectedAudioInput, stream, audioInputs } = storeToRefs(useSettingsAudioDevice())
 const chatOrchestrator = useChatOrchestratorStore()
 const chatSession = useChatSessionStore()
@@ -51,73 +52,14 @@ const sendModeLabels = computed<Record<SendMode, string>>(() => ({
   'double-enter': t('stage.send-mode.double-enter'),
 }))
 
-// Transcription pipeline
-const hearingStore = useHearingStore()
-const hearingPipeline = useHearingSpeechInputPipeline()
-const { transcribeForMediaStream, stopStreamingTranscription } = hearingPipeline
-const { supportsStreamInput } = storeToRefs(hearingPipeline)
-const { configured: hearingConfigured, autoSendEnabled, autoSendDelay } = storeToRefs(hearingStore)
-const shouldUseStreamInput = computed(() => supportsStreamInput.value && !!stream.value)
-
-// Auto-send logic
-let autoSendTimeout: ReturnType<typeof setTimeout> | undefined
-const pendingAutoSendText = ref('')
-
-function clearPendingAutoSend() {
-  if (autoSendTimeout) {
-    clearTimeout(autoSendTimeout)
-    autoSendTimeout = undefined
-  }
-  pendingAutoSendText.value = ''
-}
-
-async function debouncedAutoSend(text: string) {
-  // Double-check auto-send is enabled before proceeding
-  if (!autoSendEnabled.value) {
-    clearPendingAutoSend()
-    return
-  }
-
-  // Add text to pending buffer
-  pendingAutoSendText.value = pendingAutoSendText.value ? `${pendingAutoSendText.value} ${text}` : text
-
-  // Clear existing timeout
-  if (autoSendTimeout) {
-    clearTimeout(autoSendTimeout)
-  }
-
-  // Set new timeout
-  autoSendTimeout = setTimeout(async () => {
-    // Final check before sending - auto-send might have been disabled while waiting
-    if (!autoSendEnabled.value) {
-      clearPendingAutoSend()
-      return
-    }
-
-    const textToSend = pendingAutoSendText.value.trim()
-    if (textToSend && autoSendEnabled.value) {
-      try {
-        // `ingest()` resolves only after the full assistant turn finishes; clear UI/buffer now so
-        // the next SentenceEnd during streaming does not append to the message we already committed.
-        messageInput.value = ''
-        pendingAutoSendText.value = ''
-        const providerConfig = providersStore.getProviderConfig(activeProvider.value)
-        await ingest(textToSend, {
-          chatProvider: await providersStore.getProviderInstance(activeProvider.value) as ChatProvider,
-          model: activeModel.value,
-          providerConfig,
-        })
-      }
-      catch (err) {
-        console.error('[ChatArea] Auto-send error:', err)
-        // Preserve any transcription that arrived while ingest was in flight (see PR review).
-        messageInput.value = [textToSend, messageInput.value.trim()].filter(Boolean).join(' ')
-        pendingAutoSendText.value = [textToSend, pendingAutoSendText.value.trim()].filter(Boolean).join(' ')
-      }
-    }
-    autoSendTimeout = undefined
-  }, autoSendDelay.value)
-}
+useTranscriptions(
+  {
+    messageInputRef: messageInput,
+    sendMessage: handleSend,
+    isListeningRef: isListening,
+    isStageTamagotchi,
+  },
+)
 
 async function handleSend() {
   if (!messageInput.value.trim() || isComposing.value) {
@@ -137,7 +79,8 @@ async function handleSend() {
     })
   }
   catch (error) {
-    messageInput.value = textToSend
+    // preserve any user input when failed to send the message
+    messageInput.value = [textToSend, messageInput.value.trim()].filter(Boolean).join(' ')
     chatSession.setSessionMessages(chatSession.activeSessionId, [
       ...messages.value.slice(0, -1),
       {
@@ -229,225 +172,6 @@ watch([hearingPopoverOpen, enabled, stream], () => {
 
 onUnmounted(() => {
   teardownAnalyzer()
-  stopListening()
-
-  // Clear auto-send timeout on unmount
-  if (autoSendTimeout) {
-    clearTimeout(autoSendTimeout)
-    autoSendTimeout = undefined
-  }
-})
-
-// Transcription listening functions
-async function startListening() {
-  // Allow calling this even if already listening - transcribeForMediaStream will handle session reuse/restart
-  try {
-    console.info('[ChatArea] Starting listening...', {
-      enabled: enabled.value,
-      hasStream: !!stream.value,
-      supportsStreamInput: supportsStreamInput.value,
-      hearingConfigured: hearingConfigured.value,
-    })
-
-    // Auto-configure Web Speech API as default if no provider is configured
-    if (!hearingConfigured.value) {
-      // Check if Web Speech API is available in the browser
-      // Web Speech API is NOT available in Electron (stage-tamagotchi) - it requires Google's embedded API keys
-      // which are not available in Electron, causing it to fail at runtime
-      const isWebSpeechAvailable = typeof window !== 'undefined'
-        && !isStageTamagotchi() // Explicitly exclude Electron
-        && ('webkitSpeechRecognition' in window || 'SpeechRecognition' in window)
-
-      if (isWebSpeechAvailable) {
-        console.info('[ChatArea] No transcription provider configured. Auto-configuring Web Speech API as default...')
-
-        // Initialize the provider in the providers store first
-        try {
-          providersStore.initializeProvider('browser-web-speech-api')
-        }
-        catch (err) {
-          console.warn('[ChatArea] Error initializing Web Speech API provider:', err)
-        }
-
-        // Set as active provider
-        hearingStore.activeTranscriptionProvider = 'browser-web-speech-api'
-
-        // Wait for reactivity to update
-        await nextTick()
-
-        // Verify the provider was set correctly
-        if (hearingStore.activeTranscriptionProvider === 'browser-web-speech-api') {
-          console.info('[ChatArea] Web Speech API configured as default provider')
-          // Continue with transcription - Web Speech API is ready
-        }
-        else {
-          console.error('[ChatArea] Failed to set Web Speech API as default provider')
-          isListening.value = false
-          return
-        }
-      }
-      else {
-        console.error('[ChatArea] Web Speech API not available. No transcription provider configured and Web Speech API is not available in this browser. Please go to Settings > Modules > Hearing to configure a transcription provider. Browser support:', {
-          hasWindow: typeof window !== 'undefined',
-          hasWebkitSpeechRecognition: typeof window !== 'undefined' && 'webkitSpeechRecognition' in window,
-          hasSpeechRecognition: typeof window !== 'undefined' && 'SpeechRecognition' in window,
-        })
-        isListening.value = false
-        return
-      }
-    }
-
-    // Request microphone permission if needed (microphone should already be enabled by the user)
-    if (!stream.value) {
-      console.info('[ChatArea] Requesting microphone permission...')
-      await askPermission()
-
-      // If still no stream, try starting it manually
-      if (!stream.value && enabled.value) {
-        console.info('[ChatArea] Attempting to start stream manually...')
-        startStream()
-        // Wait for the stream to become available with a timeout.
-        try {
-          await until(stream).toBeTruthy({ timeout: 3000, throwOnTimeout: true })
-        }
-        catch {
-          console.error('[ChatArea] Timed out waiting for audio stream.')
-          isListening.value = false
-          return
-        }
-      }
-    }
-
-    if (!stream.value) {
-      const errorMsg = 'Failed to get audio stream for transcription. Please check microphone permissions and ensure a device is selected.'
-      console.error('[ChatArea]', errorMsg)
-      isListening.value = false
-      return
-    }
-
-    // Check if streaming input is supported
-    if (!shouldUseStreamInput.value) {
-      const errorMsg = 'Streaming input not supported by the selected transcription provider. Please select a provider that supports streaming (e.g., Web Speech API).'
-      console.warn('[ChatArea]', errorMsg)
-      // Clean up any existing sessions from other pages (e.g., test page) that might interfere
-      await stopStreamingTranscription(true)
-      isListening.value = false
-      return
-    }
-
-    console.info('[ChatArea] Starting streaming transcription with stream:', stream.value.id)
-
-    // Call transcribeForMediaStream - it's async so we await it
-    // Set listening state AFTER successful call
-    try {
-      await transcribeForMediaStream(stream.value, {
-        onSentenceEnd: (delta) => {
-          if (delta && delta.trim()) {
-            // Append transcribed text to message input
-            const currentText = messageInput.value.trim()
-            messageInput.value = currentText ? `${currentText} ${delta}` : delta
-            console.info('[ChatArea] Received transcription delta:', delta)
-
-            // Auto-send if enabled - check the current value (not captured in closure)
-            // This ensures we always respect the current setting, even if callbacks are reused
-            if (autoSendEnabled.value) {
-              debouncedAutoSend(delta)
-            }
-            else {
-              // If auto-send is disabled, clear any pending auto-send text to prevent accidental sends
-              clearPendingAutoSend()
-            }
-          }
-        },
-        // Omit onSpeechEnd to avoid re-adding user-deleted text; use sentence deltas only.
-      })
-
-      // Only set listening to true if transcription started successfully
-      // (transcribeForMediaStream might return early if session already exists)
-      isListening.value = true
-      console.info('[ChatArea] Streaming transcription initiated successfully')
-    }
-    catch (err) {
-      console.error('[ChatArea] Transcription error:', err)
-      isListening.value = false
-      throw err // Re-throw to be caught by outer catch
-    }
-  }
-  catch (err) {
-    console.error('[ChatArea] Failed to start transcription:', err)
-    isListening.value = false
-  }
-}
-
-async function stopListening() {
-  if (!isListening.value)
-    return
-
-  try {
-    console.info('[ChatArea] Stopping transcription...')
-
-    // Clear auto-send timeout
-    clearPendingAutoSend()
-
-    // Send any pending text immediately if auto-send is enabled
-    if (autoSendEnabled.value && pendingAutoSendText.value.trim()) {
-      const textToSend = pendingAutoSendText.value.trim()
-      pendingAutoSendText.value = ''
-      try {
-        const providerConfig = providersStore.getProviderConfig(activeProvider.value)
-        await ingest(textToSend, {
-          chatProvider: await providersStore.getProviderInstance(activeProvider.value) as ChatProvider,
-          model: activeModel.value,
-          providerConfig,
-        })
-        messageInput.value = ''
-      }
-      catch (err) {
-        console.error('[ChatArea] Auto-send error on stop:', err)
-      }
-    }
-
-    await stopStreamingTranscription(true)
-    isListening.value = false
-    console.info('[ChatArea] Transcription stopped')
-  }
-  catch (err) {
-    console.error('[ChatArea] Error stopping transcription:', err)
-    isListening.value = false
-  }
-}
-
-// Start listening when microphone is enabled and stream is available
-watch(enabled, async (val) => {
-  if (val && stream.value) {
-    // Microphone was just enabled and we have a stream, start transcription
-    await startListening()
-  }
-  else if (!val && isListening.value) {
-    // Microphone was disabled, stop transcription
-    await stopListening()
-  }
-})
-
-// Start listening when stream becomes available (if microphone is enabled)
-watch(stream, async (val) => {
-  if (val && enabled.value && !isListening.value) {
-    // Stream became available and microphone is enabled, start transcription
-    await startListening()
-  }
-  else if (!val && isListening.value) {
-    // Stream was lost, stop transcription
-    await stopListening()
-  }
-})
-
-// Watch for auto-send setting changes and clear pending sends if disabled
-watch(autoSendEnabled, (enabled) => {
-  if (!enabled) {
-    // Auto-send was disabled - clear any pending auto-send
-    clearPendingAutoSend()
-    console.info('[ChatArea] Auto-send disabled, cleared pending text')
-  }
 })
 
 watch(sendMode, () => {
